requirements.txt
google-cloud-documentai
pandas
openpyxl
python-dotenv
reportlab